Use size_t instead of int to store bit index for peers
Use size_t instead of int to store bit index for peers in RibOut,
RibOutUpdates, UpdateQueue and BgpUpdateSender.

Partial-Bug: 1602347

+++ src/bgp/bgp_ribout_updates.h
@@ -75,8 +75,8 @@ public:
                              const RibPeerSet &mready, RibPeerSet *blocked);
 
     // Enqueue a marker at the head of the queue with this bit set.
-    bool QueueJoin(int queue_id, int bit);
-    void QueueLeave(int queue_id, int bit);
+    bool QueueJoin(int queue_id, size_t bit);
+    void QueueLeave(int queue_id, size_t bit);
 
     bool Empty() const;
     size_t queue_size(int queue_id) const;
